Show task clusters when browsing a challenge.
When a user expands a challenge in the challenge result list during
challenge discovery, the locator map will be fit to the bounds of the
challenge and a bounding box will be temporarily displayed while the
clustered tasks are retrieved from the server. Once the tasks are
retrieved, the bounding box will be replaced with the task clusters.