Make challenge instructions collapsible. Fixes #29.
Users now have the ability to collapse (and re-expand) the challenge
instructions when completing tasks. This is especially useful for long
instructions, which can take up quite a bit of real-estate. The user's
selection is saved for the remainder of the challenge so that it's not
necessary to re-collapse the instructions for each task.