Add virtual challenges and map-bounded tasks
> Note that this is currently an experimental feature disabled by
> default. It can be enabled by setting the
> REACT_APP_FEATURE_BOUNDED_TASK_BROWSING .env setting to `enabled`.

With this feature enabled, the LocatorMap will now show tasks located
within the map bounds (regardless of challenge) when the map is zoomed
in enough to encompass a reasonably-small area. The maximum allowed
size of the map bounds can be controlled with the
REACT_APP_BOUNDED_TASKS_MAX_DIMENSION .env setting.

When map-bounded tasks are displayed, a new control appears in the
sidebar allowing the user to create and start a virtual challenge
consisting of the tasks visible on the map.