Avoid geojson beach-balling Safari. Fixes #222
Certain GeoJSON files take forever in Safari to be parsed and processed
by json-lint-lines, a dependency of geojsonhint (the package we use for
linting GeoJSON files). To work around that, we now parse the JSON
ourselves and pass an object to geojsonhint, side-stepping the use of
json-lint-lines. The downside is that we lose line numbers in error
messages.