Reuse task completion map. Closes #298
Reuse task completion map across tasks when completing a challenge. To
prevent potential confusion when a user moves to a new task within the
same geographic area that results in very little change to the map,
animate rendering of the task paths and markers to provide a visual cue
that a new task has been rendered on the map.