Add create-time option to append challenge URL to changeset comments.
Sometimes it's helpful to be able to trace back a changeset to the challenge that it corresponds to, and the intent and instructions given to the map contributor. This stores the option as a boolean on the Challenge object stored on the backend, and uses that flag to optionally append the URL to changeset comments.