Support viewport short-codes in comments
* Support viewport short-codes that reference a zoom/lat/lon, such as
`[v 17/37.11777/126.99754]`

* Short-codes expand to links to the referenced location on
OpenStreetMap, but on click will instead zoom JOSM to the viewport if
that is the user's configured editor

* For convenience, a viewport short-code can alternatively be created
by pasting an OSM map url, e.g.
`[https://www.openstreetmap.org/#map=17/37.11777/126.99754]`

* Supported variants are `v` or `viewport` optionally followed by a
space or slash and then slash-separated zoom, latitude, and longitude
values. For example, all of the following are equivalent:
  - `[v17/37.11777/126.99754]`
  - `[v 17/37.11777/126.99754]`
  - `[v/17/37.11777/126.99754]`
  - `[viewport17/37.11777/126.99754]`
  - `[viewport 17/37.11777/126.99754]`
  - `[viewport/17/37.11777/126.99754]`
  - `[https://www.openstreetmap.org/#map=17/37.11777/126.99754]`