When skipping a task with a status, let the server know.
We were suppressing the call to the server if someone skipped a
task that already has a status. This meant that the server would
not know that the person had skipped the task and if it was the
last task in the highest priority it would continue to be served
to the user again. (And when we get served same task back we
assume the challenge is finished for that user.)