Remove strikethrough styling on invert table filters
Custom input field for column filter was causing a weird
delay when typing where only one character would register
at a time. The purpose of the custom input field was only
to do a strike-through on invert. Removed the custom input
field and all other invert strike-throughs to be consistent.